Refactor(components): Extract blog data to separate .js file and create BlogCard component
