+++ src/components/data/blogPosts.js
@@ -1,0 +1,74 @@
+export const articleData = [
+  {
+    id: 1,
+    title: "Understanding Cat Behavior: Why Your Feline Friend Acts the Way They Do",
+    excerpt:
+      "Dive into the curious world of cat behavior, exploring why cats knead, purr, and chase imaginary prey. This article helps pet owners decode their feline's actions and understand how their instincts as hunters shape their daily routines.",
+    image: "../assets/images/blog/blog-img1.jpeg",
+    category: "Cat",
+    authorImage: "../assets/images/blog/blog-img1.jpeg",
+    authorName: "Kittiphat",
+    launchDate: "11 September 2024",
+    url: "#",
+  },
+  {
+    id: 2,
+    title: "The Fascinating World of Cats: Why We Love Our Furry Friends",
+    excerpt:
+      "Cats have captivated human hearts for thousands of years. Whether lounging in a sunny spot or playfully chasing a string, these furry companions bring warmth and joy to millions of homes. But what makes cats so special? Let's dive into the unique traits, behaviors, and quirks that make cats endlessly fascinating.",
+    image: "../assets/images/blog/blog-img1.jpeg",
+    category: "Cat",
+    authorImage: "../assets/images/blog/blog-img1.jpeg",
+    authorName: "Kittiphat",
+    launchDate: "11 September 2024",
+    url: "#",
+  },
+  {
+    id: 3,
+    title: "Finding Motivation: How to Stay Inspired Through Life's Challenges",
+    excerpt:
+      "This article explores strategies to maintain motivation when faced with personal or professional challenges. From setting small goals to practicing mindfulness and surrounding yourself with positive influences, it provides actionable tips to reignite your passion and keep moving forward.",
+    image: "../assets/images/blog/blog-img1.jpeg",
+    category: "Cat",
+    authorImage: "../assets/images/blog/blog-img1.jpeg",
+    authorName: "Kittiphat",
+    launchDate: "11 September 2024",
+    url: "#",
+  },
+  {
+    id: 4,
+    title: "The Science of the Cat’s Purr: How It Benefits Cats and Humans Alike",
+    excerpt:
+      "Discover the fascinating science behind the cat's purr, including its potential healing properties for both cats and humans. Learn how this unique sound is produced and the emotional and physical benefits it brings to both species.",
+    image: "../assets/images/blog/blog-img1.jpeg",
+    category: "Cat",
+    authorImage: "../assets/images/blog/blog-img1.jpeg",
+    authorName: "Kittiphat",
+    launchDate: "11 September 2024",
+    url: "#",
+  },
+  {
+    id: 5,
+    title: "Unlocking Creativity: Simple Habits to Spark Inspiration Daily",
+    excerpt:
+      "Discover practical ways to nurture creativity in your everyday life. Whether it's through journaling, taking breaks, or exploring new hobbies, this article offers simple yet effective habits to help you tap into inspiration and stay creatively charged.",
+    image: "../assets/images/blog/blog-img1.jpeg",
+    category: "Cat",
+    authorImage: "../assets/images/blog/blog-img1.jpeg",
+    authorName: "Kittiphat",
+    launchDate: "11 September 2024",
+    url: "#",
+  },
+  {
+    id: 6,
+    title: "Top 10 Health Tips to Keep Your Cat Happy and Healthy",
+    excerpt:
+      "This guide offers essential tips on keeping your cat in peak health. Covering topics like proper nutrition, regular vet visits, grooming, and mental stimulation, it’s a must-read for cat owners who want to ensure their pets live long, happy lives.",
+    image: "../assets/images/blog/blog-img1.jpeg",
+    category: "Cat",
+    authorImage: "../assets/images/blog/blog-img1.jpeg",
+    authorName: "Kittiphat",
+    launchDate: "11 September 2024",
+    url: "#",
+  },
+];
